Rename authentication and payment files
These were originally created with the wrong casing, and have been changed
to be consistent with the rest of the codebase.
