+++ src/ui/structure/index.ts
@@ -1,4 +1,3 @@
 export { Container } from './container'
 export { Header } from './header'
-export { WrappedCard } from './wrappedCard'
 export { CardWrapper } from './cardWrapper'
